feat: Добавлены новые функции для пользователей и админов
- Добавлена регистрация для незарегистрированных пользователей
- Добавлена кнопка 'Мои депозиты' с отображением активных депозитов
- Добавлена кнопка 'Восстановить финпароль' в главное меню
- Добавлены настройки уведомлений (депозиты, выводы, маркетинг)
- Добавлено обновление контактов (телефон, email)
- Добавлена кнопка 'Настроить уровни депозитов' в админ-панель
- Обновлены тексты кнопок в ADMIN_MENU_BUTTONS
- Исправлены все ошибки линтера (критичные и некритичные)
- Добавлена модель UserNotificationSettings с миграцией
- Созданы репозиторий и сервис для настроек уведомлений
- Добавлены FSM состояния для обновления профиля
- Улучшены сообщения для неверифицированных пользователей

+++ app/models/appeal.py
@@ -6,7 +6,7 @@ Tracks user appeals for blocked accounts.
 
 from datetime import UTC, datetime
 
-from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import DateTime, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base
@@ -53,7 +53,7 @@ class Appeal(Base):
 
     # User who submitted appeal
     user_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False
+        Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False
     )
 
     # Related blacklist entry
@@ -71,7 +71,7 @@ class Appeal(Base):
 
     # Review information (optional, set when reviewed)
     reviewed_by_admin_id: Mapped[int | None] = mapped_column(
-        BigInteger, ForeignKey("admins.id", ondelete="SET NULL"), nullable=True
+        Integer, ForeignKey("admins.id", ondelete="SET NULL"), nullable=True
     )
 
     review_notes: Mapped[str | None] = mapped_column(Text, nullable=True)
